+++ errpref/errprefdto.go
@@ -414,7 +414,7 @@ func (ePrefDto *ErrPrefixDto) GetEPrefCollectionLen() int {
 // This unsigned integer value controls the maximum character limit
 // for this specific ErrPrefixDto instance. This maximum limit will
 // remain in force for the life of this ErrPrefixDto instance or
-// until a call is made to the method 'SetMaxErrPrefTextLineLength()'
+// until a call is made to the method 'SetMaxTextLineLen()'
 // which is used to change the value.
 //
 //
@@ -1100,9 +1100,9 @@ func (ePrefDto *ErrPrefixDto) SetIsLastLineTermWithNewLine(
 		isLastLineTerminatedWithNewLine
 }
 
-// SetMaxErrPrefTextLineLength - Sets the maximum limit on the
-// number of characters allowed in an error prefix text line output
-// for display purposes.
+// SetMaxTextLineLen - Sets the maximum limit on the number of
+// characters allowed in an error prefix text line output for
+// display purposes.
 //
 // Setting this value will control the maximum character limit for
 // this specific ErrPrefixDto instance. This maximum limit will
@@ -1116,13 +1116,14 @@ func (ePrefDto *ErrPrefixDto) SetIsLastLineTermWithNewLine(
 //
 // Input Parameters
 //
-//  maxErrPrefixTextLineLength      uint
+//  maxErrPrefixTextLineLength      int
 //     - This unsigned integer value will be used to set the
 //       maximum number of characters allowed in a text display
 //       line for error prefix information.
 //
 //       If 'maxErrPrefixTextLineLength' is set to a value less
-//       than ten (10), this method will take no action and return.
+//       than ten (10) or to a value greater than one-million
+//       (1,000,000), this method will take no action and exit.
 //
 //
 // -----------------------------------------------------------------
@@ -1132,8 +1133,8 @@ func (ePrefDto *ErrPrefixDto) SetIsLastLineTermWithNewLine(
 //  --- NONE ---
 //
 //
-func (ePrefDto *ErrPrefixDto) SetMaxErrPrefTextLineLength(
-	maxErrPrefixTextLineLength uint) {
+func (ePrefDto *ErrPrefixDto) SetMaxTextLineLen(
+	maxErrPrefixTextLineLength int) {
 
 	if ePrefDto.lock == nil {
 		ePrefDto.lock = new(sync.Mutex)
@@ -1143,22 +1144,23 @@ func (ePrefDto *ErrPrefixDto) SetMaxErrPrefTextLineLength(
 
 	defer ePrefDto.lock.Unlock()
 
-	if maxErrPrefixTextLineLength < 10 {
+	if maxErrPrefixTextLineLength < 10 ||
+		maxErrPrefixTextLineLength > 1000000 {
 		return
 	}
 
 	ePrefDto.maxErrPrefixTextLineLength =
-		maxErrPrefixTextLineLength
+		uint(maxErrPrefixTextLineLength)
 }
 
-// SetMaxErrPrefTextLineLengthToDefault - Maximum Error Prefix Line
+// SetMaxTextLineLenToDefault - Maximum Error Prefix Line
 // Length is the maximum limit on the number of characters allowed
 // in a single error prefix text line.
 //
 // This method resets that maximum limit to its default value of
 // 40-characters.
 //
-func (ePrefDto *ErrPrefixDto) SetMaxErrPrefTextLineLengthToDefault() {
+func (ePrefDto *ErrPrefixDto) SetMaxTextLineLenToDefault() {
 
 	if ePrefDto.lock == nil {
 		ePrefDto.lock = new(sync.Mutex)