+++ errpref/zzzt_errpref_0002000_test.go
@@ -9,7 +9,7 @@ func TestErrPrefixDto_CopyIn_000100(t *testing.T) {
 	ePDto := ErrPrefixDto{}.NewEPrefOld(initialStr)
 
 	// Setting Line Length to 60-Characters
-	ePDto.SetMaxErrPrefTextLineLength(60)
+	ePDto.SetMaxTextLineLen(60)
 
 	ePDto.SetEPrefCtx(
 		"Tx7.TrySomethingNew()",
@@ -66,7 +66,7 @@ func TestErrPrefixDto_CopyOut_000100(t *testing.T) {
 	ePDto := ErrPrefixDto{}.NewEPrefOld(initialStr)
 
 	// Setting Line Length to 60-Characters
-	ePDto.SetMaxErrPrefTextLineLength(60)
+	ePDto.SetMaxTextLineLen(60)
 
 	ePDto.SetEPrefCtx(
 		"Tx7.TrySomethingNew()",
@@ -121,7 +121,7 @@ func TestErrPrefixDto_Multiple_000100(t *testing.T) {
 	ePDto := ErrPrefixDto{}.NewEPrefOld(initialStr)
 
 	// Setting Line Length to 60-Characters
-	ePDto.SetMaxErrPrefTextLineLength(60)
+	ePDto.SetMaxTextLineLen(60)
 
 	ePDto.SetEPrefCtx(
 		"Tx7.TrySomethingNew()",
@@ -177,7 +177,7 @@ func TestErrPrefixDto_Multiple_000100(t *testing.T) {
 			actualStr)
 	}
 
-	ePDto.SetMaxErrPrefTextLineLengthToDefault()
+	ePDto.SetMaxTextLineLenToDefault()
 }
 
 func TestErrPrefixDto_Multiple_000200(t *testing.T) {
@@ -192,7 +192,7 @@ func TestErrPrefixDto_Multiple_000200(t *testing.T) {
 	ePDto := ErrPrefixDto{}.NewEPrefOld(initialStr)
 
 	// Setting Line Length to 40-Characters
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	ePDto.SetEPrefCtx(
 		"Tx7.TrySomethingNew()",
@@ -258,7 +258,7 @@ func TestErrPrefixDto_Multiple_000200(t *testing.T) {
 			actualStr)
 	}
 
-	ePDto.SetMaxErrPrefTextLineLengthToDefault()
+	ePDto.SetMaxTextLineLenToDefault()
 }
 
 func TestErrPrefixDto_Multiple_000300(t *testing.T) {
@@ -273,7 +273,7 @@ func TestErrPrefixDto_Multiple_000300(t *testing.T) {
 	ePDto := ErrPrefixDto{}.NewEPrefOld(initialStr)
 
 	// Setting Line Length to 40-Characters
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	ePDto.SetEPrefCtx(
 		"Tx7.TrySomethingNew()",
@@ -339,7 +339,7 @@ func TestErrPrefixDto_Multiple_000300(t *testing.T) {
 			actualStr)
 	}
 
-	ePDto.SetMaxErrPrefTextLineLengthToDefault()
+	ePDto.SetMaxTextLineLenToDefault()
 }
 
 func TestErrPrefixDto_NewEPrefOld_000100(t *testing.T) {
@@ -351,7 +351,7 @@ func TestErrPrefixDto_NewEPrefOld_000100(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.NewEPrefOld(initialStr)
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	actualStr := ePDto.String()
 
@@ -382,7 +382,7 @@ func TestErrPrefixDto_NewEPrefOld_000200(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.NewEPrefOld(initialStr)
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	actualStr := ePDto.String()
 
@@ -409,7 +409,7 @@ func TestErrPrefixDto_SetCtx_000100(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	initialStr :=
 		"Tx1.AVeryVeryLongMethodNameCalledSomething() : A->B\nTx2.SomethingElse() : A==B\n" +
@@ -450,7 +450,7 @@ func TestErrPrefixDto_SetCtx_000200(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	initialStr :=
 		"Tx1.AVeryVeryLongMethodNameCalledSomething() : A->B\nTx2.SomethingElse() : A==B\n" +
@@ -491,7 +491,7 @@ func TestErrPrefixDto_SetCtx_000300(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	expectedStr := ""
 
@@ -522,7 +522,7 @@ func TestErrPrefixDto_SetEPref_000100(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	initialStr :=
 		"Tx1.AVeryVeryLongMethodNameCalledSomething() : A->B\nTx2.SomethingElse() : A==B\n" +
@@ -564,7 +564,7 @@ func TestErrPrefixDto_SetEPref_000200(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	expectedStr := "Tx5.BrandNewMethod()"
 
@@ -595,7 +595,7 @@ func TestErrPrefixDto_SetEPref_000300(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	initialStr :=
 		"Tx1.AVeryVeryLongMethodNameCalledSomething() : A->B\nTx2.SomethingElse() : A==B\n" +
@@ -636,7 +636,7 @@ func TestErrPrefixDto_SetEPref_000400(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	initialStr := "Tx1.StartThis()"
 
@@ -671,7 +671,7 @@ func TestErrPrefixDto_SetEPrefCollection000100(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	initialStr :=
 		"Tx1.Something() - Tx2.SomethingElse() - Tx3.DoSomething()\n" +
@@ -686,7 +686,7 @@ func TestErrPrefixDto_SetEPrefCollection000100(t *testing.T) {
 
 	ePDto2 := ErrPrefixDto{}.New()
 
-	ePDto2.SetMaxErrPrefTextLineLength(40)
+	ePDto2.SetMaxTextLineLen(40)
 
 	initialStr2 :=
 		"ZzA1.AVeryVeryLongMethodNameCalledSomething() : A->B\nZzA2.SomethingElse() : A==B\n" +
@@ -733,7 +733,7 @@ func TestErrPrefixDto_SetEPrefCollection000200(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	initialStr :=
 		"Tx1.Something() - Tx2.SomethingElse() - Tx3.DoSomething()\n" +
@@ -750,7 +750,7 @@ func TestErrPrefixDto_SetEPrefCollection000200(t *testing.T) {
 
 	ePDto2 := ErrPrefixDto{}.New()
 
-	ePDto2.SetMaxErrPrefTextLineLength(40)
+	ePDto2.SetMaxTextLineLen(40)
 
 	ePDto2.SetIsLastLineTermWithNewLine(true)
 
@@ -811,7 +811,7 @@ func TestErrPrefixDto_SetEPrefCtx_000100(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	initialStr :=
 		"Tx1.AVeryVeryLongMethodNameCalledSomething() : A->B\nTx2.SomethingElse() : A==B\n" +
@@ -854,7 +854,7 @@ func TestErrPrefixDto_SetEPrefCtx_000200(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	initialStr :=
 		"Tx1.AVeryVeryLongMethodNameCalledSomething() : A->B\nTx2.SomethingElse() : A==B\n" +
@@ -897,7 +897,7 @@ func TestErrPrefixDto_SetEPrefCtx_000300(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	initialStr :=
 		"Tx1.AVeryVeryLongMethodNameCalledSomething() : A->B\nTx2.SomethingElse() : A==B\n" +
@@ -939,7 +939,7 @@ func TestErrPrefixDto_SetEPrefCtx_000400(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	initialStr :=
 		"Tx1.AVeryVeryLongMethodNameCalledSomething() : A->B\nTx2.SomethingElse() : A==B\n" +
@@ -981,7 +981,7 @@ func TestErrPrefixDto_SetIsLastLineTermWithNewLine_000100(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	initialStr :=
 		"Tx1.Something()\n" +
@@ -1025,7 +1025,7 @@ func TestErrPrefixDto_SetMaxErrPrefTextLineLength_000100(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(60)
+	ePDto.SetMaxTextLineLen(60)
 
 	initialStr :=
 		"Tx1.Something()\nTx2.SomethingElse()\nTx3.DoSomething()\nTx4() - Tx5()\nTx6.DoSomethingElse()"
@@ -1059,7 +1059,7 @@ func TestErrPrefixDto_String_000100(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	initialStr :=
 		"Tx1.Something()\nTx2.SomethingElse()\nTx3.DoSomething()\nTx4() - Tx5()\nTx6.DoSomethingElse()"
@@ -1093,7 +1093,7 @@ func TestErrPrefixDto_String_000200(t *testing.T) {
 
 	ePDto := ErrPrefixDto{}.New()
 
-	ePDto.SetMaxErrPrefTextLineLength(40)
+	ePDto.SetMaxTextLineLen(40)
 
 	initialStr := "Tx1.StartThis()"
 
@@ -1183,5 +1183,5 @@ func TestErrPrefixDto_StrMaxLineLen_000200(t *testing.T) {
 			actualStr)
 	}
 
-	ePDto.SetMaxErrPrefTextLineLengthToDefault()
+	ePDto.SetMaxTextLineLenToDefault()
 }